+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- The default buffer size increased to 16MB (up from 8kB).
+
 ## [[0.2.0](https://docs.rs/disk-spinner/0.2.0/disk-spinner/)] - 2025-08-06
 
 - Initial release on crates.io.